+++ trace-seq.c
@@ -192,7 +192,7 @@ trace_seq_vprintf(struct trace_seq *s, const char *fmt, va_list args)
  */
 int trace_seq_puts(struct trace_seq *s, const char *str)
 {
-	int len;
+	size_t len;
 
 	TRACE_SEQ_CHECK_RET0(s);
 
@@ -206,7 +206,7 @@ int trace_seq_puts(struct trace_seq *s, const char *str)
 	memcpy(s->buffer + s->len, str, len);
 	s->len += len;
 
-	return len;
+	return (int)len;
 }
 
 int trace_seq_putc(struct trace_seq *s, unsigned char c)
